requirements.txt
PySocks==1.7.1
python-telegram-bot==12.5.1
requests==2.23.0
python-dotenv==0.12.0